move get_colormap to utilities
